I got an idea to temporary solve the total won bests values, which is to inflate alertDialog s after the 2nd match in every played best from the three, And let the user tell the App who won the second match so:
- if left won 1st and 2nd matches then:
- reset recent best matches to 0 (( To start new best "the 2nd in this case ))

But this will make the if/else statements inside both + Buttons TOO LONGER than it's now.

-----------------------------------------------------------------
And for the hided menu on the shown now toolBar I'll look to the code when i come back from my bicycle ride to see why the text is overlayed " As replaced the about with the app icon and showed on the toolBar.

-----------------------------------------------------------------
BTW the Number 3 is not the winning score point, it's 21
3 is just for testing & I really wish to see anyone who know how to play Ping Pong from the mentors here.

-----------------------------------------------------------------
I'll go for my ride now, and wish to come back with the whole idea to put in the code OR seeing the right code to a method that use the updated global Integers values and put the if/else statements inside it.